Refactor naming and package to drools-retediagram as agreed
