--- src/test/java/net/tarilabs/model/Cheese.java
@@ -1,18 +1,0 @@
-package net.tarilabs.model;
-
-
-public class Cheese {
-
-    private final String name;
-
-    public Cheese(String name) {
-        super();
-        this.name = name;
-    }
-
-    
-    public String getName() {
-        return name;
-    }
-    
-}
